feat: export error contract helpers
